Experiment with premature streaming.
If the Boxee player is fed a partially downloaded file with buffered
duration T, playback will stop once T is reached, even if the download
process continues in the background. A workaround for this is to reload
the file once download has completed, and seek to where playback was
interrupted. The GUI will however flicker for some seconds when this is
done.

This patch is not really meant to be used in the initial versions (if
ever), but is kept for reference.