Changed design to use xml.svtplay.se and Flash.
Changed foundation of Boxee app. Information is retrieved from
http://xml.svtplay.se (when possible, .html files will be scraped when
no .xml interface has been found or is available). Playback relies on
the Boxee browser and SVTPlay's Flash player.

Caveats: this commit allows the first 20 items of the alphabetical index
be displayed (navigating later items is not yet implemented). When an
item is selected, playback immediately begins (a list of episodes should
be shown).

This commit renders several files obsolete, notably the server
directory.